aprgaudio refactoring

+++ AprgAudio/AprgAudio/tst/AudioTests/Audio_unit.cpp
@@ -38,12 +38,11 @@ TEST(AiffLoadingTests, Stereo_8bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_8bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_8bit_44100::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_8bit_44100::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_8bit_44100::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_8bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_16bit_44100) {
     Audio<double> audioFile;
@@ -56,12 +55,11 @@ TEST(AiffLoadingTests, Stereo_16bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_16bit_44100::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_16bit_44100::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_16bit_44100::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_16bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_24bit_44100) {
     Audio<double> audioFile;
@@ -74,12 +72,11 @@ TEST(AiffLoadingTests, Stereo_24bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_24bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_24bit_44100::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_24bit_44100::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_24bit_44100::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_24bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_8bit_48000) {
     Audio<double> audioFile;
@@ -92,12 +89,11 @@ TEST(AiffLoadingTests, Stereo_8bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_8bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_8bit_48000::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_8bit_48000::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_8bit_48000::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_8bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_16bit_48000) {
     Audio<double> audioFile;
@@ -110,12 +106,11 @@ TEST(AiffLoadingTests, Stereo_16bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_16bit_48000::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_16bit_48000::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_16bit_48000::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_16bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(AiffLoadingTests, Stereo_24bit_48000) {
     Audio<double> audioFile;
@@ -128,12 +123,11 @@ TEST(AiffLoadingTests, Stereo_24bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), aiff_stereo_24bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), aiff_stereo_24bit_48000::numChannels);
 
-    for (unsigned int i = 0; i < aiff_stereo_24bit_48000::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(aiff_stereo_24bit_48000::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], aiff_stereo_24bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WritingTest, WriteFromCopiedSampleBuffer) {
     Audio<float> audioFile1, audioFile2;
@@ -157,12 +151,11 @@ TEST(WavLoadingTests, Stereo_8bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_8bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_8bit_44100::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_8bit_44100::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_8bit_44100::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_16bit_44100) {
     Audio<double> audioFile;
@@ -175,12 +168,11 @@ TEST(WavLoadingTests, Stereo_16bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_16bit_44100::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_16bit_44100::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_16bit_44100::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_24bit_44100) {
     Audio<double> audioFile;
@@ -193,12 +185,11 @@ TEST(WavLoadingTests, Stereo_24bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_24bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_24bit_44100::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_24bit_44100::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_24bit_44100::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_44100::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Mono_16bit_44100) {
     Audio<double> audioFile;
@@ -211,12 +202,11 @@ TEST(WavLoadingTests, Mono_16bit_44100) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_mono_16bit_44100::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_mono_16bit_44100::numChannels);
 
-    for (unsigned int i = 0; i < wav_mono_16bit_44100::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_mono_16bit_44100::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_mono_16bit_44100::testBuffer[i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_8bit_48000) {
     Audio<double> audioFile;
@@ -229,12 +219,11 @@ TEST(WavLoadingTests, Stereo_8bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_8bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_8bit_48000::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_8bit_48000::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_8bit_48000::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_8bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_16bit_48000) {
     Audio<double> audioFile;
@@ -247,12 +236,11 @@ TEST(WavLoadingTests, Stereo_16bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_16bit_48000::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_16bit_48000::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_16bit_48000::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_16bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Stereo_24bit_48000) {
     Audio<double> audioFile;
@@ -265,12 +253,11 @@ TEST(WavLoadingTests, Stereo_24bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_stereo_24bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_stereo_24bit_48000::testBuffer.size());
 
-    for (unsigned int i = 0; i < wav_stereo_24bit_48000::testBuffer[0].size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_stereo_24bit_48000::testBuffer[0].size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_stereo_24bit_48000::testBuffer[k][i]));
         }
-    }
-}
+    }}
 
 TEST(WavLoadingTests, Mono_16bit_48000) {
     Audio<double> audioFile;
@@ -283,12 +270,11 @@ TEST(WavLoadingTests, Mono_16bit_48000) {
     EXPECT_EQ(audioFile.getSampleRate(), wav_mono_16bit_48000::sampleRate);
     EXPECT_EQ(audioFile.getNumChannels(), wav_mono_16bit_48000::numChannels);
 
-    for (unsigned int i = 0; i < wav_mono_16bit_48000::testBuffer.size(); i++) {
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++) {
+    for (int i = 0; i < static_cast<int>(wav_mono_16bit_48000::testBuffer.size()); i++) {
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) {
             EXPECT_TRUE(isAlmostEqual(audioFile.samples[k][i], wav_mono_16bit_48000::testBuffer[i]));
         }
-    }
-}
+    }}
 
 bool writeTest(int numChannels, int sampleRate, int bitDepth, AudioFormat format) {
     float sampleRateAsFloat = (float)sampleRate;
@@ -297,15 +283,13 @@ bool writeTest(int numChannels, int sampleRate, int bitDepth, AudioFormat format
 
     audioFile.setAudioBufferSize(numChannels, sampleRate * 4);
 
-    for (unsigned int i = 0; i < static_cast<unsigned int>(audioFile.getNumSamplesPerChannel()); i++) {
+    for (int i = 0; i < static_cast<int>(audioFile.getNumSamplesPerChannel()); i++) {
         float sample = static_cast<float>(sin(2.0 * getPi() * i / sampleRateAsFloat * 440.0));
 
-        for (unsigned int k = 0; k < static_cast<unsigned int>(audioFile.getNumChannels()); k++)
-            audioFile.samples[k][i] = sample * 0.5;
+        for (int k = 0; k < static_cast<int>(audioFile.getNumChannels()); k++) audioFile.samples[k][i] = sample * 0.5;
     }
 
-    audioFile.setSampleRate(sampleRate);
-    audioFile.setBitDepth(bitDepth);
+    audioFile.setSampleRate(sampleRate);    audioFile.setBitDepth(bitDepth);
 
     std::string numChannelsAsString = numChannels == 1 ? "mono" : "stereo";
     std::string bitDepthAsString = std::to_string(bitDepth);
